perf: Reduce unnecessary component re-render
Performance issues:
Every action that affects a block element causes the entire editor to re-render (through a different key).

Current solution：
If an Enter/Backspace KeyDown event is received that affects a block element, a predefined procedure is used instead.

Using a native event action will create an element that is not controlled by React.

# This does not apply for now when the selection type is Range or Delete KeyDown event.